Revert all changes in Comb/HW
