+++ lib/Dialect/Comb/CMakeLists.txt
@@ -17,7 +17,6 @@ add_circt_dialect_library(CIRCTComb
 
   LINK_LIBS PUBLIC
   CIRCTHW
-  CIRCTSVExternal
   MLIRIR
   MLIRInferTypeOpInterface
    )
