+++ lib/Dialect/HW/CMakeLists.txt
@@ -21,7 +21,6 @@ add_circt_dialect_library(
   Support
   
   LINK_LIBS PUBLIC
-  CIRCTSVExternal
   MLIRIR
 )
