--- lib/Dialect/SV/External/CMakeLists.txt
@@ -1,8 +1,0 @@
-add_circt_library(CIRCTSVExternal
-  SVAttributes.cpp
-
-  DEPENDS
-
-  LINK_LIBS PUBLIC
-  MLIRIR
-)
